network-scanner/requirements.txt
python-nmap==0.7.1
requests==2.31.0
schedule==1.1.0
tzlocal==5.2
netifaces==0.11.0 